APIGOV-32516 - refactor agent
